Suppress pop-up of console window during client initialization.
Issue: On Windows, a console window briefly pops up when initializing
the client from inside a GUI application. (This is not noticeable if
called from a console application, as the console window would then
simply be inherited and thus *not* pop up.)

Analysis: This is due to a known bug, ["Hide command prompt when
using `subprocess.Popen` with `shell=False` on Windows"][bug], that
has yet to be fixed as of Python 3.8. It occurs when the `backend`
module invokes `subprocess` in order to query the Comsol server's
version.

Solution: On Windows, pass `creationflags=0x08000000` to the
`subprocess.run` call, as recommended in the answer to the Stack
Overflow question ["How do I hide the console when I use `os.system()`
or `subprocess.call()`?"][so].

Note: The explicit constant (`0x08000000`) is passed for reasons of
backward-compatibility with Python 3.6 (and possibly earlier). As of
Python 3.7, it could be referred to via the pre-defined name
`subprocess.CREATE_NO_WINDOW`.

[bug]: https://bugs.python.org/issue30082
[so]:  https://stackoverflow.com/a/7006424