Added test for catching keyboard interrupts to sandbox.
The script only uses JPype directly, not MPh. It tests
an unresolved issue that arises then the user presses
Ctrl+C while the Comsol client is running, if only idling
in the background. Pressing Ctrl+C seems to crash the
Java Virtual Machine *and* the Python session as well.
This is unexpected and undesired behavior.