Unpeg JPype and Python version. Work around shutdown bug.
For versions of JPype after 0.7.5, the Java VM does not
shut down properly at the end of the Python session.
(More precisely, only after a Comsol client has been
initialized and a model been loaded. See the `test_client`
script in the `sandbox` folder for more details).

Unfortunately, JPype 0.7.5 does not work on Pyton 3.8.4
and later. In order to support the latest Python versions,
the dependency on the old JPype version was lifted. The
work-around is to deliberately crash the Java VM at the
end of the Python session. This is not clean, but so be it.