Refactored back-end discovery on Windows and Linux.
Restructured the code that discovers installed Comsol back-ends. This
makes the `client` and `server` modules independent of the operating
system - at this point at least.

This means the basic Linux compatibility proposed in pull request #3
is now available out of the box. It is limited to client-server mode.
Users need to start a server instance and then a client that connects
to it.

Stand-alone clients do not work correctly at this point. There are
issues during initialization (see test script in `sandbox` folder)
and additional issues further down the line, such as while solving
the model or when importing external image data. These issues are
currently unresolved.

Added support for MacOS as well, though it is completely untested.
It is assumed that things would basically work the same way as in
Linux, except that the install locations were changed to the folders
indicated in the Comsol documentation.