Refactored environment setup for stand-alone clients.
As detailed in issue #8, setting the environment variable
for the search path of shared libraries has no effect on
the current process on Unix-like systems. And on Windows
is not even needed. The code that does this was therefore
removed. But a check was added for Linux and macOS that tests
if the environment variable has been correctly set by the user
beforehand if, and only if, a stand-alone client is to be
initialized. Updated the documentation accordingly.