Use client-server mode in test suite from now on.
As detailed in issue #8, starting a stand-alone client does
not work out of the box on Unix-like operating systems.
Relying on the client-server mode instead seems to be the
way forward. The test suite now uses that mode by default.
Stand-alone clients can still be tested by passing the
command-line argument "stand-alone" to a test script.