Start server with "comsol mphserver" on Linux and macOS.
The command-line option "server", which was used in the code,
is deprecated as of Comsol 5.1. The option's name was changed
in favor of "mphserver" so as to avoid confusion with Comsol
Server, which was then introduced as a separate product and
does not refer to the server application that is part of Comsol
Multiphysics.

This happens to clarify the question raised in #2 as to how far
back the support for older Comsol versions goes.