Use auto-login when starting server process.
We now start the server executable with the command-line option
`-login auto`. This means that if login details (user name and
password) have not been defined previously, they will be created
automatically. Otherwise the existing login details will be used.

The benefit of this is that MPh will work out of the box, whether
or not the user has ever started the server manually. We also no
longer have to check for that particular error scenario. If, at
some point, the user does decide to define a login, they must run
the server executable with the option `-passwd reset`