Return parameters as dictionary instead of list of named tuples.
This is a breaking API change. It is noted as such in the docs,
specifically in the doc-string of `Model.parameters()`. The method
is more commonly used for interactive inspection at the Python
prompt, so this change will probably not break much in terms of
application code. The dictionary as the new return type is in line
with other new methods, such as `Model.properties()`.