Fixed failing evaluation of complex-valued expressions.
This has only been tested manually with a model from Comsol's
Application Gallery, "Frequency Domain Modeling of a Capacitor",
where the admittance is a frequency-dependent series of complex
numbers. Complex-valued return types from evaluations have yet
to be integrated into the test suite.

This might fix part of the issue reported in #35, even though
the error message there is different from the one the test model
produced before the fix.