Added demo script that generates the tutorial model.
At some point, this demo script may be used to "bootstrap"
the tutorial model entirely from Python code. Its main purpose,
however, is to demonstrate the creation of advanced features.
It is intended to be updated as new library functionality is
introduced that would simplify the syntax further.

Currently, the tags the code generates are not "canonical",
i.e. different from the tags that the Comsol GUI would create.
This is often not a big deal, as the tags are mostly hidden
from application code, but may be a problem for things like
physics interfaces, where the tag also serves as the name-space
prefix for variables used in expressions throughout the model.