﻿Improve performance by ignoring Java objects in `Node` constructor.
When creating a new `Node` object, its `__init__()` constructor
would access the Java objects of all built-in groups in order to
store them in an internal dictionary. Instead, the objects are now
stored as string references and their evaluation is deferred to the
`.java` property, i.e. only occurs when one of the objects is actually
needed.

This reduces communication overhead in client–server mode, where it
improves performance by a factor of 2 or so when navigating nodes in
the model tree. It makes little difference in stand-alone mode.