Only test file import/export with absolute paths.
Testing with relative file paths does make sense, but is tricky to
get right. It depends on whether the model was loaded from a file,
or created from scratch but hasn't been saved yet. And it may then
even depend on whether or not the model has been solved.

Running the tests directly via `run_all.py` or indirectly via pyTest
may introduce further complications. Or be subject to platform
differences.

This way, the tests in `test_model.py` can be run in source order of
the `model` module, avoiding all of the above complications.