Fixed pyTest failures on Windows.
On Windows, but not on Linux, starting the Java VM via JPype crashed
when running the test suite via pyTest. (Or rather, when running
individual test scripts from the suite, since the suite as a whole
cannot be run via pyTest anyway as the individual scripts would then
try to start the Java VM more than once, which JPype does not allow.)

The crash can be avoided by turning off Python's fault handlers, which
pyTest happens to enable, right before starting the Java VM. See section
"Errors reported by Python fault handler" in the JPype User Guide for
background. See issue jpype-project/jpype#886 for why this might be a
problem on Windows in particular. It seems to be related to how the
SIGSEGV signal is handled by Python on that platform.

On Linux, turning off the fault handlers is not necessary to make the
test suite compatible with pyTest. However, on Windows and Linux both,
errors also that occur at the very end of pyTest's run. (These may be
related to issues with JPype's shutdown of the JVM, but, regardless,
they do not affect the outcome of the tests.) On Linux with fault
handlers left on, the errors are reported as segmentation faults.
With fault handlers off, like they now are on Windows, the segmentation
faults would turn into fatal errors as "detected by the Java Runtime
Environment" and are then dumped to `hs_err_pid*.log` files, which is
even more annoying.