Added reporting of test code coverage.
Code coverage by the test suite is measured via pyTest and its
pytest-cov plug-in. As pyTest cannot run the entire test suite
in one go (as the Java VM is started multiple times, by separate
test scripts), the script `coverage.py` is provided in the `deploy`
folder to automate the runs of inidividual test scripts and combine
the reports into a unified file.

The script also creates the coverage badge, unless that file already
exists. In other words, delete the badge to update the coverage
percentage.

Test coverage is reported for Windows only. Code that targets other
platforms is therefore exluded from the reporting.