Suppress error if preference `checkforrecoveries` is undefined.
It appears this preference is not always defined, like on certain
Linux distributions or installation on Windows using a classkit
license. See issue #39. We just suppress the error as the whole
point is to not have recovery files anyway, seeing as how they might
use up all disk space when many simulations are run back to back.