Removed exit handling for exceptions raised in threads.
When a Python program exits due to an unhandled exception in a
thread other than the main thread, the exit code, as it turns out,
is actually zero. So this is maybe unexpected, but standard behavior.
There is thus no reason to try and implement a better exit handling
here, which was only done in anticipation of the hard exit of the
Java VM and was never, it seems, actually functional anyway.