Avoid code quality alerts in `Client.__init__()`.
LGTM doesn't like the no-op assignments (like `self.port = self.port`)
that we did before. And admittedly, that's ugly. It was only done so
Sphinx would render the doc-strings in the intended source-code order,
which it wouldn't when the doc-strings were put in a more natural place,
earlier in the method. This might be a bug in Sphinx. But we can structure
the code differently, and then have our cake and eat it too.