Don't stop local server instance during clean-up.
Since `mph.start()` does not start the local server instance with the
`multi=True` option, we don't have to stop it explicitly, it will just
wind down on its own. And we don't need to add another race condition
to the exit sequence, that are plenty enough as it is.