Added new logo.
© 2021 John Hennig

This logo is NOT subject to the MIT license that the library's source
code is published under. It is to be used exclusively in order to promote
this very open-source project: MPh, a "Pythonic scripting interface for
Comsol Multiphysics", currently (as of October 2021) hosted on GitHub at
https://github.com/MPh-py/MPh. Usage in any other context, namely for
derived work, requires the express permission of the copyright holder.