Fixed failure of exit-behavior tests during coverage reporting.
Just like in the recently revamped test runner (`test.py`), the project
root folder must be added to the Python search path. While pyTest does
add the root folder to `sys.path` in its own process, it is not available
in the subprocesses testing the exit behavior. Which is why these tests
failed when `tools/coverage.py` was run.