Relaxed test assertions when comparing tags with reference model.
`test_tag()` in `test_node.py` compares the node tags generated by MPh
with the tags created by Comsol for the same model. The reference model
is the demo model from the Tutorial. It was created with Comsol 5.5 and
then saved to a file, stored in the `demos` folder. Newer Comsol versions
(such as 6.0, see #67) may create default nodes that older versions didn't.
To future-proof the test, we therefore only compare the tags if the node
already existed in the reference model, but ignore "new" nodes (if any)
possibly introduced by later Comsol versions.