Added support for export of animations. (#43)
Animations can now be exported via `Model.export()`, just like images
and data files. The animation type is deduced from the file ending
as either a GIF, Flash, AVI, or WebM encoded movie, or a PNG image
sequence. This covers all animation types supported by Comsol 5.6.

Much like in #66, we no longer allow specifying nodes via a string
such as `'export/image'`. This was never documented anyway and may lead
to problems with node names that actually contain a forward slash
(which would be escaped as `//`). The node must be given as either a
string such as `'image'` or a reference like `model/'exports'/'image'`.