Generalized regular expression parsing out server port.
This should fix issue #72, where the server would not start when
Comsol was installed with the interface language set to Chinese.

Parsing is now handled by a separate function so that it can be
unit-tested. Currently, English (the default), German (issue #24),
and Chinese (issue #72) output from the server process is being
run against the regular expression. It is expected to work for
all languages, effectively just looking for a four- or five-digit
number in the line that begins with "COMSOL".