Refactored discovery mechanism to include executable search path.
The discovery mechanism was refactored. We are now first looking for
Comsol executables, i.e. `comsol.exe` on Windows or `comsol` on Linux
and macOS. We can then add to the list of candidates the Comsol
executable that the shell would find, if any, i.e. what would be called
when running `comsol` in the terminal.

This makes discovery more flexible, since that last installation does
not necessarily have to be in a default location. So finding it requires
no extra configuration other than "adding Comsol to PATH", which the
Comsol installer offers as an option (named not exactly like that, but
something to that effect).

So when there is but this one Comsol installation, MPh should work out
of the box, without the need to create symbolic links if that
installation is in a non-default location. This helps, for example,
with deployments on a cluster, where Comsol is often made available with
something like `module load comsol/6.0`.

We retain the possibility of selecting a specific Comsol back-end, for
the purpose of running tests/benchmark across versions. If the
back-end on the executable search path (a.k.a. the `PATH`) is the same
as one of those already found by searching the system (file system on
Linux and macOS, the Registry on Windows), then we'll ignore it.

To achieve this functionality, we are now starting or search for other
Comsol components from the location of the Comsol executable, i.e.
`comsol.exe` or the `comsol` shell script. We then look for `comsol.ini`
in the same folder, as it contains the relative path to the Java VM.