Add `problems()` method to `Model` class (#91)
The `Model` class gets a new API method `problems()`. All it does is
call the `problems()` method of its root node. This gives users an
easy way to check if there are any problems reported throughout the
model, by testing `if model.problems():` in application code, to then
act accordingly.