Changed Comsol version in all examples from "5.3a" to "6.0".
The example value for the `version` parameter was "5.3a" to make it
clear that it could contain letters to indicate the minor version,
as per Comsol's (old) versioning scheme. It may however be confusing
to users who would encounter it in the documention, as 5.3a is not
supported by MPh and seems to no longer work with it (although it
did in the beginning).