Don't create artifacts in `tools` folder.
This is just cosmetic. All build and most test artifacts are now created
in the `build` folder, which is the conventional location. Tools where
the output location cannot be configured, such as Coverage.py, still
produce artifacts in the root folder. But other than `__pycache__`
folders, nothing is hidden further down in the source tree. Whereas
before we put some things into the `tools` folder.