Added tooling script that invokes the code linter.
This is just for consistency. So that all dev tools are obvious just
from looking into the `tooling` folder. For now, this particular script
is super simple, as it just calls Flake8. But we might add a type
checker going forward, which would then be called as part of the linting
step.