Change error message for when no Comsol installation was found.
The old error message ("Could not locate any Comsol installation.") was
often misunderstood to mean that MPh somehow failed to find any
installed Comsol whatsoever, when in fact it did, but discarded it for
one reason or another.

This usually happens if the folder structure of the Comsol installation
is different from what we expect. Whether it's an old Comsol version or
a new CPU architecture. We don't even want to try and start those
unfamiliar installations.

Relates to issue #174.