Change error message for when users try to instantiate a second client.
The old error message ("Only one client can be instantiated at a time.")
was misunderstood to mean that if users hard-deleted the Python client
object (with `del client`), they could hope to work around this
limitation. The documentation already has plenty of information on that,
but the error message is clearer if we once again emphasize that the
limitation applies to the entire Python session.

Relates to discussion #173.