Upgrade custom Sphinx extensions for rendering Markdown doc-strings.
The new version monkey-patches Sphinx's Autodoc somewhat less and is
therefore less prone to breakage due to upstream changes. Not that
we had this problem here, just pulling in this version from a different
project where certain things wouldn't work with the old version. And
that's almost certainly still the case for whatever Sphinx features
we've never event tested against.