Assume all dependencies are installed when building docs.
This means the setup is as described in "Local development" in
`tools/ReadMe.md`, though possibly only with the `[doc]` extras
installed instead of all `[dev]` extras. The build setup on
Read-the-Docs already works like that, so we would not expect any
problems there.

In the Sphinx configuration, we then no longer need to manipulate the
Python path (`sys.path`) to be able to `import mph`, and also don't
have to mock any of the dependencies.