Fixed rendering of API documentation.
This was due to a change in Sphinx 8.2 that broke our custom MyST
extension to get Markdown support in doc-strings. It was not due to a
change in MyST itself, as suspected in the last commit. MyST happens
to restrict the Sphinx version, which is why our pin on MyST also pinned
Sphinx.

Essentially, Sphinx's Autodoc extension now has a new internal method
`Documenter._generate()`, in addition to the existing `.generate()`.
And in some cases it calls the internal method directly, bypassing
the existing method, as what it does is redundant. This broke our hack
as we then didn't get a chance to close the Sphinx domain directives we
had opened.

Also fixed some typos, polished doc-strings and code comments, added
type hints to method signatures.