Make all meta information static.
First and foremost, this means the version number, which is now in
`pyproject.toml` and no dynamically retrieved from in `mph/__init__.py`
by the build backend (Flit).

This is often considered good practice, but has a downside:
`mph.__version__` can only return the version number if the package has
been installed. This will not work when running directly from source.
We do however still want users to be able to do just that, so they can
run the test suite right after downloading the source code without much
hassle.

For that reason, we drop `test_meta.py` from the test suite. It never
tested anything that was part of the public API anyway.