Added support for ARM-based Apple Silicon architecture on macOS.
This required a bit of refactoring in the discovery code. We now detect
the platform architecture right upfront and use it to inform our
expectation of the name of Comsol's binary folder, such as `win64` or
`macarm64`. We don't expect it to behave any different on platforms that
were previously already supported.