Prioritize Comsol installation found on search path.
There may be situations where multiple installations of the same Comsol
version exist on the same machine, but the (current) user wants to pick
the one which has a single-user license in their name. The easiest way
to do that is to just change that user's PATH variable once and for all.

Related discussion: #172