Relaxed Sphinx requirement to `>8` (from `>8.2`) to make `uv sync` work.
I don't remember why I required `Sphinx>= 8.2` in the first place, but
it may have had something to do with rendering the type annotations
in the API documentation. Which I've given up on… for now.

More importantly, though, it stops `uv sync` from resolving the
dependencies, as Sphinx 8.2+ does not support Python 3.10. But Sphinx
8.0 does.