Made output of log messages the default when running test suite.
Before, we had to run `tools/test.py --log` to have it display the
debug-level log messages produced by MPh, which is helpful when
trouble-shooting problems reported by users. This behavior is now the
default. The detailed log can be suppressed by passing `--quiet` to get
the old behavior. It does make it easier to see if all tests have passed
at a glance, and how long it all took.