Make client–server mode the default (also) on Windows.
Some users reported errors on Windows that I couldn't reproduce, but
that could be or could possibly be resolved by running in client–server
mode. Which has been the default on Linux and macOS since the beginning,
but not on Windows, where we've tried to leverage the better performance
(in certain, rare scenarios) of a stand-alone client.

We now make client–server mode the default on all platforms. To get
`mph.start()` to return a stand-alone client, set `mph.option('session',
'stand-alone')` up front. This usually only makes sense on Windows,
where stand-alone clients (typically) work out of the box. You could
also set `mph.option('session', 'platform-dependent')` to get the
stand-alone client only on Windows, but not when running on other
platforms. That's the old behavior. But this option may be deprecated at
some point (though it isn't yet), as it makes little sense to offer it.

Related issues and discussions: #224 and #225.