Removed `rewrite_properties()` from `test_node.py`.
This test would read the properties of all nodes of the demo model and
write them back, i.e. get the value and re-assign it for all properties
in the model. Because of that, it made a lot of calls to the Comsol API.
That's slow in client–server mode, which is why this test was only ever
run on Windows, where we had the stand-alone client. As that's no longer
the case, this part of the test suite doesn't run, ever. So we may as
well remove it.