Require `uv_build>=0.9.8` to fix capitalization of project name on PyPI.
A bug in the `uv_build` build backend (astral-sh/uv#16547) incorrectly
lowercased the package name from "MPh" to "mph" in the metadata of the
distribution wheel. This led to the package being spelled differently
on PyPI than it used to be. Package name resolution is of course case-
insensitive, so this is a cosmetic change. But we want the metadata of
the published package to be correct.